add reviewer model logic
