+++ project/models/__init__.py
@@ -8,7 +8,7 @@ Comment:
  
 Have a good code time!
 -----
-Last Modified: Thursday January 9th 2025 12:29:05 pm
+Last Modified: Friday January 10th 2025 9:08:20 am
 Modified By: the developer formerly known as Kaixu Chen at <chenkaixusan@gmail.com>
 -----
 HISTORY:
@@ -20,3 +20,4 @@ Date 	By 	Comments
 '''
 
 from project.models.make_model import *
+from project.models.optical_flow import *
